requirements.txt
azure-mgmt-resource
azure-mgmt-compute
azure-mgmt-network
azure-identity
azure-loganalytics
azure-monitor-query
azure-storage-blob